Store the reference to the deleted entry as uint32
An earlier version of this library used int64, and this one line
stayed in place after it was changed to uin32. Thanks to aletheia7
for finding this bug.

+++ tree.go
@@ -39,7 +39,7 @@ func (tree *Critbit) addExternalRef(key string, value interface{}) (uint32, erro
 func (tree *Critbit) deleteExternalRef(refNum uint32) {
 	tree.numExternalRefs--
 	tree.externalRefs[refNum].key = ""
-	tree.externalRefs[refNum].value = int64(tree.firstDeletedRef)
+	tree.externalRefs[refNum].value = uint32(tree.firstDeletedRef)
 	tree.firstDeletedRef = refNum
 }
